requirements.txt
requests~=2.27.1
Pillow~=8.4.0